+++ backend/src/delivery/domain/status_record.rs
@@ -4,14 +4,14 @@ use crate::delivery::domain::value_objects::ParcelStatus;
 
 #[derive(Clone)]
 pub struct StatusRecord {
-    pub id: usize,
-    pub parcel_id: usize,
+    pub id: i32,
+    pub parcel_id: i32,
     pub status: ParcelStatus,
     pub creation_time: DateTime<Utc>,
 }
 
 impl StatusRecord {
-    pub fn new(parcel_id: usize, status: ParcelStatus) -> Self {
+    pub fn new(parcel_id: i32, status: ParcelStatus) -> Self {
         StatusRecord {
             id: 0,
             parcel_id,
